+++ core/model/src/main/kotlin/ru/resodostudios/cashsense/core/model/data/Wallet.kt
@@ -2,11 +2,12 @@ package ru.resodostudios.cashsense.core.model.data
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
 data class Wallet(
     val id: String,
     val title: String,
-    val startBalance: Double,
+    val startBalance: BigDecimal,
     val currency: Currency
 ) : Parcelable
