+++ feature/transaction/src/main/kotlin/ru/resodostudios/cashsense/feature/transaction/TransactionDialog.kt
@@ -95,7 +95,7 @@ fun AddTransactionDialog(
                         walletOwnerId = walletId,
                         categoryId = category.id,
                         description = description,
-                        amount = amount.toDouble(),
+                        amount = amount.toBigDecimal(),
                         date = Clock.System.now()
                     )
                 )
@@ -190,7 +190,7 @@ fun EditTransactionDialog(
                         walletOwnerId = transactionWithCategory.transaction.walletOwnerId,
                         categoryId = category?.id,
                         description = description,
-                        amount = amount.toDouble(),
+                        amount = amount.toBigDecimal(),
                         date = transactionWithCategory.transaction.date
                     )
                 )
